requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
django-crispy-forms==2.4
djangorestframework==3.16.0
pdfkit==1.0.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.2
python-dotenv
